requirements.txt
flaky
locust
IPy
alembic==1.1.0
allure-pytest==2.8.6
allure-python-commons==2.8.6
apipkg==1.5
astroid==2.3.3
async-generator==1.10
async-lru==1.0.2
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
colorama==0.4.1
coverage==5.0.3
cryptography==2.8
distro==1.4.0
docopt==0.6.2
enum34==1.1.6
et-xmlfile==1.0.1
execnet==1.7.1
idna==2.8
importlib-metadata==1.5.0
isort==4.3.21
jdcal==1.4.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
namedlist==1.7
oauthlib==3.1.0
openpyxl==3.0.3
packaging==19.2
paramiko==2.6.0
pipreqs==0.4.9
pluggy==0.13.0
ply==3.11
py==1.8.0
pycparser==2.19
pylint==2.4.3
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.2
pyrsistent==0.15.7
pytest==5.2.1
pytest-cov==2.7.1
pytest-forked==1.1.3
pytest-html==2.0.0
pytest-metadata==1.8.0
pytest-parallel==0.0.9
pytest-xdist==1.30.0
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.22.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.12.0
SQLAlchemy==1.3.8
thrift==0.11.0
thriftapi===stable
thriftpy2==0.4.8
tornado==6.0.2
urllib3==1.25.6
wcwidth==0.1.7
wrapt==1.11.2
yarg==0.1.9
zipp==3.0.0
